net/mlx5e: Fix parsing of vlan packets when updating lro header
Previous to this patch vlan tagged packets were not parsed correctly
and assumed to be regular IPv4/IPv6 packets.
We should check for 802.1Q/802.1ad tags and update the lro header
accordingly.
This fixes the use case where LRO is on and rxvlan is off
(vlan stripping is off).

Issue: 852921
Change-Id: I4c32fc45f72265c2827b3d057228a80e1171171b
Fixes: e586b3b0baee ('net/mlx5: Ethernet Datapath files')
Signed-off-by: Gal Pressman <galp@mellanox.com>