Cover letter: mlx5_ib RoCE LAG
This feature keeps track, using netdev LAG events,
of the bonding and link status of each port's PF netdev.

When both of the card's PF netdevs are enslaved exclusively
to the same bond master, LAG state is active.
Bond mode must be one of:
1 (active-backup), 2 (XOR), or 4 (802.3ad).

During LAG, a single IB device is present for both ports.
This allows for load balancing and high availablity to be
managed by the driver, abstracting it away from ULPs.
Such an IB device is given a name of "mlx5_bond_X".

LAG mode is determined by the bond driver.
In load balancing (modes 2 and 4), QPs are assigned to ports
in a round-robin fashion, on QP transition from RESET->INIT.
In high availability (mode 1), all QPs are assigned to the active
slave, determined by bond driver.

Please note that if the link state of a port becomes down when
mode is load balancing, all QPs will be moved to the other port,
and will be moved back once both ports are up again.

This patchset is split into mlx5_core and mlx5_ib parts:
1) mlx5_core: Implements RoCE LAG infrastructure and API.
   a) Bond device LAG events tracking.
   b) Decisions as to whether LAG state is active.
   c) FW commands to enter and exit LAG state.
   d) Manages HA behavior.
2) mlx5_ib: Uses core infrastructure to implement RoCE LAG.
   a) Chooses an "mlx5_bond_X" device name when LAG is active.
   b) Implements QP load balancing.
   c) Merges steering, such that both port's traffic arrives on
      PF0's root flow table. This is used for processing
      both port's usermode Ethernet traffic on PF0.
   d) Creates a special flow table that diverts all
      non-usermode Ethernet traffic received on port 2
      back to PF1's root flow table (since non-usermode Ethernet
      should not be affected by LAG).

Issue: 746988
Change-Id: I40c76e06ac8c3a870a77815d88638b66f6dd2658
Signed-off-by: Aviv Heller <avivh@mellanox.com>