net/mlx5e: Add basic TC tunnel set action for SRIOV offloads
The code maintains mlx5_encap_entry for each encap header
it has encountered when trying to offload TC tunnle set action.
This entry maintains a linked list of all the flows sharing the same
encap header, when the last flow is removed from the list the encap
entry is removed.

The actual encap_header is allocated in the hardware only if we have
layer 2 neighbour info when mlx5_encap_entry is created.
The field valid_encap tells us whether the header is in the hardware and encap_id
holds the encap_id that was allocated.
While the flow is in the driver, the driver holds a reference on the neighbour.

When a new flow with encap action is inserted the code first checks
if the required encap entry exist according to the tunnel set parameters,
if it does the encap is shared, otherwise a new mlx5_encap_entry is created.

The flow is inserted to the FDB only if the corresponding encap is
valid. If the encap is invalid we don't insert the flow to the FDB and
we don't keep it in the driver.

TC action parsing implementation in the driver assumes that tunnel set
action is parsed before mirred_redirect action.

Issue: 840526
Change-Id: I98b01e4aa3e174d204df37183c37b885fd41779c
Signed-off-by: Hadar Hen Zion <hadarh@mellanox.com>
Signed-off-by: Saeed Mahameed <saeedm@mellanox.com>