IB/mlx4: Handle well-known-gid in mad_demux processing
If OpenSM runs over ConnectX-3, and there are ConnectIB VFs active
on the network, the OpenSM will receive QP1 packets containing a GRH
where the destination GID is the "Well-Known GID" -- which is not a GID
in the HCA Port's GID Table.

This GID must be tested-for separately -- and packets which contain
this destination GID should be routed to slave 0 (the PF).

Fixes: 37bfc7c1e83f ('IB/mlx4: SR-IOV multiplex and demultiplex MADs')
Signed-off-by: Jack Morgenstein <jackm@dev.mellanox.co.il>
Signed-off-by: Daniel Jurgens <danielj@mellanox.com>
Signed-off-by: Leon Romanovsky <leon@kernel.org>