IB/mlx5: Modify mlx5_ib to support 4K UARs for userspace
Modify the implementation such that we allocate the required number of
UARs depending on firmware support of 4K UARs, libmlx5 support and the
number of blue flame registers required by the user. We set requested
library support to false in this patch. A subsequent patch will add
support conveying this requirement from libmlx5.

Signed-off-by: Eli Cohen <eli@mellanox.com>
Reviewed-by: Matan Barak <matanb@mellanox.com>
Signed-off-by: Leon Romanovsky <leon@kernel.org>