IB/ipoib: Add warning message when changing the MTU
In datagram mode, the IB UD (Unreliable Datagram) transport is used
so the MTU of the interface is equal to the IB L2 MTU minus the
IPoIB encapsulation header.

Any request to change the MTU value above the maximum range will change
the MTU to the max allowed, but will not show any warning message.

An ipoib_warn is issued in such cases, letting the user know that even
though the value is legal, but it can't be currently applied.

Signed-off-by: Feras Daoud <ferasda@mellanox.com>
Signed-off-by: Noa Osherovich <noaos@mellanox.com>
Signed-off-by: Leon Romanovsky <leon@kernel.org>