IB/ipoib: Disable events over child interface that is in the middle of deletion
The flow of deleting a child interface, goes through netdevice un-registration
and calls to flush the ipoib workqueue and cancel all the works in the queue.
If the queue contains a heavy_flush work, it will trigger all the works again
in ipoib_ib_dev_open function and they will run over the interface that is
going to be destroyed. This will lead to unexpected results.
The fix is to mark the interface as IPOIB_FLAG_GOING_DOWN, and not trigger
events for it in such cases.

Signed-off-by: Feras Daoud <ferasda@mellanox.com>
Signed-off-by: Erez Shitrit <erezsh@mellanox.com>
Signed-off-by: Leon Romanovsky <leon@kernel.org>