IB/ipoib: Fix race issue with delayed work on child kill flow
This fix address the following issue:
flow 1:
ipoib_reap_neigh(struct work_struct *work)
1-1)    if (!test_bit(IPOIB_STOP_NEIGH_GC, &priv->flags))
1-2)    queue_delayed_work(priv->wq, &priv->neigh_reap_task,
                           arp_tbl.gc_interval)

flow 2:
ipoib_neigh_hash_uninit(struct net_device *dev)
2-1)	test_and_set_bit(IPOIB_STOP_NEIGH_GC, &priv->flags)
2-2)	cancel_delayed_work(&priv->neigh_reap_task)

The Linux Kernel is preemptable, which means that this sequence of
actions is feasible.
	(1-1)
	(2-1)
	(2-2)
	(1-2)

In this case, (2-2) of this sequence has no effect, the work
will re-schedule itself and run one more time.
In IPoIB, after each "set and cancel" sequence, flush is called,
so during the work run, the flush starts and panic occur.
To fix this, cancel_delayed_work_sync should be used, by this
before reaching the flush, we will make sure that the work is done.

Signed-off-by: Feras Daoud <ferasda@mellanox.com>
Signed-off-by: Erez Shitrit <erezsh@mellanox.com>
Signed-off-by: Leon Romanovsky <leon@kernel.org>