net/mlx5e: Remove PCIe counters timers and states group
This group is yet to have a capability bit, as a result users with
old firmware will syndrome every 200ms due to update stats work.
Will be pushed again in the future with the required bit.
Note that other PCIe groups will be queried using a capability introduced
in the next patches.

Issue: 908955
Issue: 907593
Change-Id: I173788f294a78501968f176dff2aeda6e1e2fe7c
Fixes: 9c7262399ba1 ("net/mlx5e: Expose PCIe statistics to ethtool")
Reported-by: Jesper Dangaard Brouer <brouer@redhat.com>
Signed-off-by: Gal Pressman <galp@mellanox.com>