net/mlx5e: Remove xdpsq->db.doorbell
It is not needed since we can sample XDP tx producer index at the start
of RQ NAPI cycle and check if it was changed in the process, if yes
then a doorbell is required.

Advance stats.xdp_tx only once at the end of such cycle using the same
sampled XDP tx producer.

Issue: 894740
Change-Id: I138b0196e7eb9ec94ce11179cbbd7ab8bb437309
Signed-off-by: Saeed Mahameed <saeedm@mellanox.com>