IB/core: Add DEVICE type and root types structure
This adds the DEVICE type. This type supports creating the context that
all objects are created from. Moreover, it supports executing actions
which are related to the device itself, such as QUERY_DEVICE.
The only related object to this type is a singleton (per file
instance) context.

All standard types are put in the root structure. This root will later
on be used in drivers as their parsing tree root. Later on, when new
features are added, these drivers could define their own root.

Signed-off-by: Matan Barak <matanb@mellanox.com>
Reviewed-by: Yishai Hadas <yishaih@mellanox.com>
Signed-off-by: Leon Romanovsky <leon@kernel.org>