net/mlx5e: Properly handle misc errors while adding TC rules
When the FW returns an error (common example is an attempt to offload
twice the same rule which is refused by some FWs), we are not
properly derefing/cleaning some resources allocated on the way.

Examples are vport vlan deref under eswitch vlan offloads, and encap
entry / neighbour deref under eswitch encapsulation offloads, fix that.

Fixes: a54e20b4fcae ('net/mlx5e: Add basic TC tunnel set action for SRIOV offloads')
Fixes: 8b32580df1cb ('net/mlx5e: Add TC vlan action for SRIOV offloads')
Signed-off-by: Or Gerlitz <ogerlitz@mellanox.com>
Signed-off-by: Saeed Mahameed <saeedm@mellanox.com>