net/mlx5e: Support TC encapsulation offloads with upper devices
When tunneling is used, it's common in virtualization systems that the
mlx5e uplink device may be stacked under upper devices such as bridge or
ovs internal port, where the VTEP IP address used for the encapsulation
is set on that upper device.

In order to support that case, we use eswitch service function returning
the uplink device and set it as the mirred device of the tc encap rule.

We remove the check for the HW ID since it doesn't apply to all use
cases and can fail us, for example when mlx5 uplink is stacked under ovs
internal bridge.

Issue: 912137
Change-Id: I397cc53f98fca5d6c3e102531a37b38c5877a512
Fixes: a54e20b4fcae ("net/mlx5e: Add basic TC tunnel set action for SRIOV offloads")
Signed-off-by: Hadar Hen Zion <hadarh@mellanox.com>
Reviewed-by: Or Gerlitz <ogerlitz@mellanox.com>