net/mlx5: Enlarge the NIC TC offload table size
The NIC TC offload table size was hard coded to 1k. Change it to be

      min(max NIC RX table size, max flow counters * num flow groups)

where the max values are read from the firmware and the number of
flow groups is hard-coded as before this change.

We don't know upfront the division of flows to group. This setup allows
each group to be of size up to the where we want to support (when supported,
all offloaded flows use counters). Thus, we don't expect multiple occurences
for a group which in turn adds steering hops.

issue: 900706
Change-Id: I137e6cb41637ba9b3208a6683712541d69ca0fa4
Signed-off-by: Or Gerlitz <ogerlitz@mellanox.com>